reorganization into modules
