--- app/src/main/java/ch/epfl/skysync/dataModels/AvailabilityCalendar.kt
@@ -1,13 +1,0 @@
-package ch.epfl.skysync.dataModels
-
-import java.time.LocalDate
-
-class AvailabilityCalendar: CalendarModel<Availability>() {
-    fun setAvailabilityByDate(date: LocalDate, timeSlot: TimeSlot, status: AvailabilityStatus) {
-        setByDate(date, timeSlot) { d, t , old -> old.setStatus(status)}
-    }
-
-    fun getAvailabilityByDate(date: LocalDate, timeSlot: TimeSlot): Availability? {
-        return getByDate(date, timeSlot)
-    }
-}
