--- app/src/main/java/ch/epfl/skysync/dataModels/TimeSlot.kt
@@ -1,5 +1,0 @@
-package ch.epfl.skysync.dataModels
-
-enum class TimeSlot {
-    AM, PM
-}
