+++ app/src/main/java/ch/epfl/skysync/dataModels/calendarModels/TimeSlot.kt
@@ -1,0 +1,5 @@
+package ch.epfl.skysync.dataModels.calendarModels
+
+enum class TimeSlot {
+    AM, PM
+}
