+++ app/src/main/java/ch/epfl/skysync/dataModels/flightModels/Balloon.kt
@@ -1,4 +1,4 @@
-package ch.epfl.skysync.dataModels
+package ch.epfl.skysync.dataModels.flightModels
 
 data class Balloon(val name:String, val qualification: BalloonQualification)
