+++ app/src/main/java/ch/epfl/skysync/dataModels/flightModels/FlightType.kt
@@ -1,4 +1,4 @@
-package ch.epfl.skysync.dataModels
+package ch.epfl.skysync.dataModels.flightModels
 
 data class FlightType(
     val name: String,
