+++ app/src/main/java/ch/epfl/skysync/dataModels/flightModels/Role.kt
@@ -1,4 +1,7 @@
-package ch.epfl.skysync.dataModels
+package ch.epfl.skysync.dataModels.flightModels
+
+import ch.epfl.skysync.dataModels.flightModels.RoleType
+import ch.epfl.skysync.dataModels.userModels.User
 
 class Role(
     val roleType: RoleType,
