+++ app/src/main/java/ch/epfl/skysync/dataModels/flightModels/RoleType.kt
@@ -1,4 +1,4 @@
-package ch.epfl.skysync.dataModels
+package ch.epfl.skysync.dataModels.flightModels
 
 enum class RoleType {
     PILOT,
