+++ app/src/main/java/ch/epfl/skysync/dataModels/flightModels/Team.kt
@@ -1,4 +1,4 @@
-package ch.epfl.skysync.dataModels
+package ch.epfl.skysync.dataModels.flightModels
 
 class Team {
     val roles = mutableListOf<Role>()
